test(arrayFilter): move example to examples_test file
